Move ConstantValues impl to metal::library
